Removed no longer needed userdata.cpp
